requirements.txt
fastapi
uvicorn
boto3
python-multipart
pypdf
python-dotenv
requests
playwright
